modeling/requirements.txt
click==7.1.2
rq
PyYAML>=5.3
intelhex
angr==8.19.10.30
IPython
pyyaml
